+++ src/LASR/lasralgorithm.h
@@ -37,6 +37,7 @@ public:
   virtual bool process(LASpoint*& p) { return true; };
   virtual bool process(LAS*& las) { return true; };
   virtual bool process(LAScatalog*& las) { return true; };
+  virtual bool process() { return true; };
   virtual bool write() { return true; };
   virtual void clear(bool last = false) { return; };
   virtual bool set_chunk(const Chunk& chunk);